+++ frontend/src/App.jsx
@@ -4,7 +4,7 @@ import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-d
 import ProtectedRoute from './components/protectedRoute';
 import Login from './pages/Login';
 import Register from './pages/Register';
-import CalendarDashboard from './pages/CalendarDashboard';
+import Dashboard from './pages/Dashboard';
 import AuthService from './services/authenticationService';
 
 function App() {
@@ -13,14 +13,14 @@ function App() {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={isAuthenticated ? <CalendarDashboard /> : <Login />} />
+        <Route path="/" element={isAuthenticated ? <Dashboard /> : <Login />} />
         <Route path="/login" element={isAuthenticated ? <Navigate to = "/home"/> : <Login />} />
         <Route path="/register" element={<Register />} />
 
         {/*Protected routes*/}
         <Route element={<ProtectedRoute />}>
           {/* <Route path="/home" element={<HomePage />} /> */}
-          <Route path="/home" element={<CalendarDashboard />} />
+          <Route path="/home" element={<Dashboard />} />
         </Route>
         {/* <Route path="/home" element={<ProtectedRoute><HomePage /></ProtectedRoute>} /> */}
       </Routes>
