+++ frontend/src/styles/Calendar.css
@@ -1,0 +1,67 @@
+
+.calendar-arrows {
+    display: flex;
+    gap: 30px;
+    font-size: 20px;
+    margin: 5px 0px 15px 0px;
+  
+    .calendar-year, .calendar-month{
+      margin: 0 4px;
+    }
+  
+    .calendar-arrow-btn {
+      background: none;
+      border: none;
+      font-size: 22px;
+      cursor: pointer;
+      border: 1px solid rgb(51, 51, 51);
+      border-radius: 50%;
+    }
+  }
+  
+  .calendar-grid-container {
+    display: flex;
+    justify-content: center;
+    align-items: flex-start;
+    margin-top: 10px;
+    height: 100%
+  }
+  
+  .calendar-table {
+    border-spacing: 0;
+    width: 100%;
+    height: 95%;
+    border: 1px solid #222;
+    overflow: hidden;
+  }
+  
+  .calendar-header-cell {
+    border: 1px solid #222;
+    padding: 10px 0;
+    font-size: 18px;
+    font-weight: bold;
+  }
+  
+  .calendar-cell {
+    position: relative;
+    border: 1px solid #222;
+    min-height: 60px;
+    min-width: 60px;
+  }
+  
+  .calendar-date-number {
+    position: absolute;
+    top: 6px;
+    left: 50%;
+    transform: translateX(-50%);
+    font-size: 14px;
+    font-weight: bold;
+    padding: 2px;
+  }
+  
+  .calendar-date-number.today {
+    background: black;
+    color: white;
+    border-radius: 50%;
+    padding: 4px 10px;
+  }
