+++ frontend/src/styles/Dashboard.css
@@ -1,28 +1,28 @@
 body {
   margin: 0px;
 }
-.calendar-container {
+.page-container {
   display: flex;
   height: 100vh;
   width: 100vw;
   font-family: Arial, sans-serif;
 }
 
-.calendar-sidebar {
+.page-sidebar {
   width: 220px;
   border-right: 2px solid #222;
   display: flex;
   flex-direction: column;
 }
 
-.calendar-profile {
+.profile {
   display: flex;
   align-items: center;
   height: 60px;
   padding: 10px 16px;
   border-bottom: 1px solid grey;
 
-  .calendar-profile-picture {
+  .profile-picture {
     width: 50px;
     height: 50px;
     border-radius: 50%;
@@ -30,21 +30,21 @@ body {
     margin-right: 12px;
   }
 
-  .calendar-profile-name {
+  .profile-name {
     font-weight: bold;
     font-size: 18px;
     letter-spacing: 1.5px;
   }
 }
 
-.calendar-tabs-list {
+.sidebar-tabs {
   margin-top: 24px;
   display: flex;
   flex-direction: column;
   gap: 25px;
   padding-left: 24px;
 
-  .calendar-tab {
+  .new-tab {
     font-size: 18px;
     cursor: pointer;
     transition: color 0.2s;
@@ -56,13 +56,14 @@ body {
   }
 }
 
-.calendar-main {
+.main-page {
   flex: 1;
   display: flex;
   flex-direction: column;
   padding: 18px;
 
-  .calendar-header {
+  .page-header {
+    flex: 1 1 1;
     display: flex;
     justify-content: left;
     align-items: center;
@@ -81,6 +82,21 @@ body {
     }
   }
 
+  .header-action {
+    cursor: pointer;
+  }
+}
+
+.main-content {
+  height: 90%;
+}
+
+.calendar-arrows {
+  display: flex;
+  gap: 30px;
+  font-size: 20px;
+  margin: 5px 0px 15px 0px;
+
   .calendar-year, .calendar-month{
     margin: 0 4px;
   }
@@ -93,18 +109,14 @@ body {
     border: 1px solid rgb(51, 51, 51);
     border-radius: 50%;
   }
-
-  .calendar-header-action {
-    cursor: pointer;
-  }
 }
 
 .calendar-grid-container {
-  flex: 1;
   display: flex;
   justify-content: center;
   align-items: flex-start;
   margin-top: 10px;
+  height: 100%
 }
 
 .calendar-table {
